+++ lib/core/app/runtimeDatas.dart
@@ -1,6 +1,6 @@
 import 'package:animestream/core/data/types.dart';
 import 'package:animestream/core/database/anilist/types.dart';
-import 'package:animestream/ui/models/subtitles.dart';
+import 'package:animestream/ui/models/widgets/subtitles.dart';
 import 'package:animestream/ui/theme/types.dart';
 import 'package:flutter_dotenv/flutter_dotenv.dart';