+++ lib/ui/models/widgets/subtitles.dart
@@ -1,6 +1,7 @@
 // ignore_for_file: public_member_api_docs, sort_constructors_first
 import 'dart:convert';
 
+import 'package:animestream/core/commons/enums.dart';
 import 'package:animestream/core/data/preferences.dart';
 import 'package:animestream/ui/models/extensions.dart';
 import 'package:better_player/src/video_player/video_player.dart';
@@ -8,8 +9,6 @@ import 'package:flutter/material.dart';
 
 import 'package:animestream/core/commons/subtitleParsers.dart';
 import 'package:animestream/ui/models/snackBar.dart';
-
-enum SubtitleFormat { ASS, VTT }
 
 class Subtitle {
   final Duration start;