+++ lib/ui/pages/lists.dart
@@ -2,7 +2,7 @@ import 'package:animestream/core/app/runtimeDatas.dart';
 import 'package:animestream/core/commons/enums.dart';
 import 'package:animestream/core/database/anilist/queries.dart';
 import 'package:animestream/core/database/anilist/types.dart';
-import 'package:animestream/ui/models/cards.dart';
+import 'package:animestream/ui/models/widgets/cards.dart';
 import 'package:flutter/material.dart';
 
 class AnimeLists extends StatefulWidget {