+++ lib/ui/pages/settingPages/player.dart
@@ -1,7 +1,7 @@
 import 'package:animestream/core/app/runtimeDatas.dart';
 import 'package:animestream/core/data/settings.dart';
 import 'package:animestream/core/data/types.dart';
-import 'package:animestream/ui/models/slider.dart';
+import 'package:animestream/ui/models/widgets/slider.dart';
 import 'package:animestream/ui/pages/settingPages/common.dart';
 import 'package:animestream/ui/pages/settingPages/subtitle.dart';
 import 'package:flutter/material.dart';